Adding event cleaning option for cutflow studies

+++ EasyjetHub/EasyjetHub/CutManager.h
@@ -20,10 +20,10 @@ struct CutEntry
 {
     std::string name{}; // Name of the cut
     int counter{0}; //Counts the events passed by the cut.
-    float w_counter{0.f}; //Counts the events passed by the cut in terms of mcEventWeights.at(0)
+    double w_counter{0.f}; //Counts the events passed by the cut in terms of mcEventWeights.at(0)
     bool passed{0}; // If event is passed this should be set to true. Otherwise, it should be set to false.
     int relativeCounter{0}; // Counts the events passed with respect to the previous cuts. 
-    float w_relativeCounter{0.f}; // Counts the events passed with respect to the previous cuts in terms of mcEventWeights.at(0)
+    double w_relativeCounter{0.f}; // Counts the events passed with respect to the previous cuts in terms of mcEventWeights.at(0)
     
     CutEntry(const std::string& n)
         : name(n) {}
